backend/requirements.txt
groq>=0.11.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
langchain-chroma>=0.1.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0

fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
